+++ src/toggle_group.rs
@@ -41,7 +41,7 @@ impl SelectOneOfState {
 
 pub fn input_toggle_group(e: &mut GuiEvent) -> Option<OutputAction> {
     match &e.event_type {
-        InputEvent::Tap(pt) => {
+        InputEvent::PointerUp(pt) => {
             e.scene.mark_dirty_view(e.target);
             e.scene.set_focused(e.target);
             if let Some(view) = e.scene.get_view_mut(e.target) {